Connect pizza and ingredient backend to frontend

+++ lib/data/domain/pizza_create_request.dart
@@ -1,13 +1,13 @@
 import 'package:pizza_app/data/domain/ingredient.dart';
 
 class PizzaCreateRequest {
-  final int price;
+  final double price;
   final String name;
   final List<Ingredient> ingredients;
 
-  PizzaCreateRequest(
-      {required this.price,
-      required this.name,
-      required this.ingredients,
-      });
+  PizzaCreateRequest({
+    required this.price,
+    required this.name,
+    required this.ingredients,
+  });
 }
